scale/src/main/java/com/strategyobject/substrateclient/scale/readers/I16Reader.java
package com.strategyobject.substrateclient.scale.readers;

import com.strategyobject.substrateclient.common.streams.StreamUtils;
import com.strategyobject.substrateclient.scale.ScaleReader;
import lombok.NonNull;
import lombok.val;

import java.io.IOException;
import java.io.InputStream;
import java.nio.ByteBuffer;
import java.nio.ByteOrder;

public class I16Reader implements ScaleReader<Short> {
    @Override
    public Short read(@NonNull InputStream stream) throws IOException {
        val bytes = StreamUtils.readBytes(2, stream);
        ByteBuffer buf = ByteBuffer.allocate(2).order(ByteOrder.LITTLE_ENDIAN);
        buf.put(bytes);
        buf.flip();
        return buf.getShort();
    }
}
